Extract memberlist.HTTPStatusHandler
Instead of having the http handler logic in the KVInitService, we now
have it in a separate entity that is instantiated with a default
template.
